Fix lint warnings (#37)
* Fix lint warnings

Mostly, fix some small typos.

Rename 'dot' to 'dotFile' since 'dot' shadows an imported package of the
same name. Since we are renaming 'dot', also rename the corresponding
'href' variable so they are of consistent form.

Rename 'MermaidShapeCirle' to 'MermaidShapeCircle' as the former was
likely a typo.

Verified by running 'go test ./...' and confirming all tests passed.

* go mod tidy

* Fix Go formatting

Run 'go fmt' (go version 1.21.6) to verify that all files are formatted properly.

* Mark MermaidShapeCirle as deprecated

Instead of removing MermaidShapeCirle, mark it as deprecated. Separate
it from the test for the other Mermaid shapes so that when the variable
is eventuallly removed from the codebase it will be trivial to also
remove its test.